user_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
sqlalchemy
python-dotenv
python-multipart
typing-extensions
PyJWT
aiosqlite
pytest
starlette
gunicorn
